requirements.txt
pandas
scikit-learn
GitPython
joblib
flask

